requirements.txt
selenium==4.16.0
webdriver-manager==4.1.0
python-dotenv==1.0.0
pdfplumber==0.9.0
pdfminer.six==20221105
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
